Add files via upload
In this project we aim to analyze the South St. Petersburg Community Redevelopment Area (CRA) that
was established in 2015 to promote reinvestment in housing and neighborhoods, non-residential
properties, education, and workforce development in the city. We aim to help the city with these findings
by understanding historical trends of property values since the adoption of the CRA. This should help
leadership to make actionable decisions for future investments in the communities.